moved endpoints.rs to separate files
